Implemented fraud detection engine with three rules: large amounts, balance drain, rapid transactions

+++ src/main/java/com/transactguard/transactguard/repo/FraudRepository.java
@@ -1,7 +1,8 @@
 package com.transactguard.transactguard.repo;
 
 import com.transactguard.transactguard.entity.FraudFlag;
+import com.transactguard.transactguard.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FraudRepository extends JpaRepository<FraudFlag, Integer> {
+public interface FraudRepository extends JpaRepository<FraudFlag, Long> {
 }
